fix: change writing logic and add option to keep input file

+++ rq_geo_toolkit/geoparquet_sorting.py
@@ -38,6 +38,7 @@ def sort_geoparquet_file_by_geometry(
     row_group_size: int = PARQUET_ROW_GROUP_SIZE,
     working_directory: Union[str, Path] = "files",
     verbosity_mode: "VERBOSITY_MODE" = "transient",
+    remove_input_file: bool = True,
 ) -> Path:
     """
     Sorts a GeoParquet file by the geometry column.
@@ -65,6 +66,8 @@ def sort_geoparquet_file_by_geometry(
             verbosity mode. Can be one of: silent, transient and verbose. Silent disables
             output completely. Transient tracks progress, but removes output after finished.
             Verbose leaves all progress outputs in the stdout. Defaults to "transient".
+        remove_input_file (bool, optional): Remove the original file after sorting.
+            Defaults to True.
     """
     if output_file_path is None:
         output_file_path = (
@@ -95,7 +98,8 @@ def sort_geoparquet_file_by_geometry(
             pq.read_metadata(input_file_path)
         )
 
-        input_file_path.unlink()
+        if remove_input_file:
+            input_file_path.unlink()
 
         order_files = sorted(order_dir_path.glob("*.parquet"), key=lambda x: int(x.stem))
 
@@ -305,19 +309,14 @@ def _order_single_row_group(
         for rg_id, local_rows_ids in indexes_to_read_per_row_group.items()
     }
 
-    schema = pq.read_schema(original_file_path)
-
-    pq.write_table(
-        table=pa.concat_tables(
-            [
-                # Read rows from each read row group using reshuffled local indexes
-                read_tables_per_row_group[rg_id].take(reshuffled_indexes)
-                for rg_id, reshuffled_indexes in reshuffled_indexes_to_read
-            ]
-        ),
-        where=output_dir_path / f"{row_group_id}.parquet",
-        schema=schema,
+    # Read rows from each read row group using reshuffled local indexes
+    concatenated_tables = pa.concat_tables(
+        [
+            read_tables_per_row_group[rg_id].take(reshuffled_indexes)
+            for rg_id, reshuffled_indexes in reshuffled_indexes_to_read
+        ]
     )
+    pq.write_table(table=concatenated_tables, where=output_dir_path / f"{row_group_id}.parquet")
 
 
 def _calculate_row_group_mapping(file_path: Path) -> dict[int, tuple[int, int]]:
